refactor(rag-knowledge,pipeline-controller): U3 — Pipeline 系 3 Port + Adapter 新設、Controller を委譲化 (#730)
Issue #730 の Unit 3。`PipelineController` (1197 LOC) を 3 つの責務単位
Port + Adapter に分離し、Controller を薄い委譲層に縮小する。

新 Port + Adapter:

- `src/rag/pipeline/change_detector.py`
  - `ChangeDetector` Protocol + `RealChangeDetector` 実装
  - `scan_all_as_added` / `supplement_hidden_changes` / `classify_changes`
  - `_map_status` を private モジュール関数に切り出し
- `src/rag/pipeline/metadata_builder.py`
  - `MetadataBuilder` Protocol + `RealMetadataBuilder` 実装
  - `register_in_db` / `update_in_db` / `resolve_source_id` /
    `build_metadata` / `build_metadata_from_record`
  - `_read_meta_dict` は private 内部ヘルパー
- `src/rag/pipeline/change_handler.py`
  - `ChangeHandler` Protocol + `RealChangeHandler` 実装
  - `process_change` / `_handle_added` / `_handle_modified` /
    `_handle_deleted` / `_handle_renamed` / `_handle_meta_only`

`PipelineController` の縮小:

- `__init__` で 3 Adapter を構築
- `_scan_all_as_added` / `_supplement_hidden_changes` / `_classify_changes` /
  `_map_status` を ChangeDetector に委譲
- `_register_in_db` / `_update_in_db` / `_resolve_source_id` /
  `_build_metadata` / `_build_metadata_from_record` を MetadataBuilder に委譲
- `_process_single_change` / `_handle_*` を ChangeHandler に委譲
- LOC: 1197 → 879（約 27% 削減）
- 既存テストの内部メソッド呼び出し（`ctrl._classify_changes` 等）は proxy
  メソッドが委譲することで互換維持

`_run_processing_loop` / `_build_error_entry` /
`_check_uncommitted_changes` / `_auto_commit_for_incremental` は引き続き
Controller 内部の補助メソッドとして残置（パイプライン実行のオーケストレーション
責務に内包される）。

検証:
- ruff / mypy 0 error（src/rag 130+ source files）
- pytest 全 2026 件 pass
